Use size_t as count type in VulkanCppConsumerBase::BuildValue
Because it is always safe to convert uint32_t to size_t while the reverse isn't true,
modify BuildValue() to take size_t's for the count values when passing in arrays.

+++ framework/decode/vulkan_cpp_consumer_base.h
@@ -611,11 +611,11 @@ class VulkanCppConsumerBase : public VulkanConsumer
     static std::string BuildValue(const StdVideoAV1FrameRestorationType* values, uint32_t count);
 
     template <typename T, class = typename std::enable_if<std::is_arithmetic<T>::value>::type>
-    static std::string BuildValue(const T* values, uint32_t count)
+    static std::string BuildValue(const T* values, size_t count)
     {
         std::stringstream output;
         output << "{";
-        for (uint32_t idx = 0; idx < count; idx++)
+        for (size_t idx = 0; idx < count; idx++)
         {
             output << std::to_string(values[idx]) << ", ";
         }
@@ -624,14 +624,14 @@ class VulkanCppConsumerBase : public VulkanConsumer
     }
 
     template <typename T, class = typename std::enable_if<std::is_arithmetic<T>::value>::type>
-    static std::string BuildValue(const T* values, uint32_t count1, uint32_t count2)
+    static std::string BuildValue(const T* values, size_t count1, size_t count2)
     {
         std::stringstream output;
         output << "{";
-        for (uint32_t idx1 = 0; idx1 < count1; idx1++)
+        for (size_t idx1 = 0; idx1 < count1; idx1++)
         {
             output << "{";
-            for (uint32_t idx2 = 0; idx2 < count2; idx2++)
+            for (size_t idx2 = 0; idx2 < count2; idx2++)
             {
                 output << std::to_string(values[(idx1 * count2) + idx2]) << ", ";
             }
